Dolphin moved to hostile
