+++ src/pocketmine/entity/Entity.php
@@ -44,6 +44,7 @@ use pocketmine\entity\hostile\WitherSkeleton;
 use pocketmine\entity\hostile\Zombie;
 use pocketmine\entity\hostile\ZombiePigman;
 use pocketmine\entity\hostile\ZombieVillager;
+use pocketmine\entity\hostile\Dolphin;
 use pocketmine\entity\object\ArmorStand;
 use pocketmine\entity\object\ExperienceOrb;
 use pocketmine\entity\object\FallingBlock;
@@ -56,7 +57,6 @@ use pocketmine\entity\object\ItemEntity;
 use pocketmine\entity\passive\Cat;
 use pocketmine\entity\passive\Chicken;
 use pocketmine\entity\passive\Cod;
-use pocketmine\entity\passive\Dolphin;
 use pocketmine\entity\passive\Pufferfish;
 use pocketmine\entity\passive\Salmon;
 use pocketmine\entity\passive\Cow;
